+++ swifter/swifter.py
@@ -3,6 +3,7 @@ import warnings
 import numpy as np
 import pandas as pd
 
+from abc import abstractmethod
 from math import ceil
 from psutil import cpu_count
 from dask import dataframe as dd
@@ -12,8 +13,14 @@ from os import devnull
 from tqdm.auto import tqdm
 from .tqdm_dask_progressbar import TQDMDaskProgressBar
 
-from numba.core.errors import TypingError
+ERRORS_TO_HANDLE = [AttributeError, ValueError, TypeError, KeyError]
+try:
+    from numba.core.errors import TypingError
 
+    ERRORS_TO_HANDLE.append(TypingError)
+except ImportError:
+    pass
+ERRORS_TO_HANDLE = tuple(ERRORS_TO_HANDLE)
 warnings.filterwarnings("ignore", category=FutureWarning)
 
 SAMPLE_SIZE = 1000
@@ -48,7 +55,7 @@ class _SwifterObject:
                 "This pandas object has duplicate indices, and swifter may not be able to improve performance. Consider resetting the indices with `df.reset_index(drop=True)`."
             )
         self._nrows = self._obj.shape[0]
-        self._SAMPLE_SIZE = SAMPLE_SIZE if self._nrows > 25000 else int(ceil(self._nrows / 25))
+        self._SAMPLE_SIZE = SAMPLE_SIZE if self._nrows > (25 * SAMPLE_SIZE) else int(ceil(self._nrows / 25))
 
         if npartitions is None:
             self._npartitions = cpu_count() * 2
@@ -205,7 +212,7 @@ class SeriesAccessor(_SwifterObject):
                     .map_partitions(func, *args, meta=meta, **kwds)
                     .compute(scheduler=self._scheduler)
                 )
-        except (AttributeError, ValueError, TypeError, KeyError):
+        except ERRORS_TO_HANDLE:
             # if map partitions doesn't match pandas apply, we can use dask apply, but it will be a bit slower
             if self._progress_bar:
                 with TQDMDaskProgressBar(desc=self._progress_bar_desc or "Dask Apply"):
@@ -247,13 +254,7 @@ class SeriesAccessor(_SwifterObject):
                     error_message="Vectorized function sample doesn't match pandas apply sample.",
                 )
             return func(self._obj, *args, **kwds)
-        except (
-            AttributeError,
-            ValueError,
-            TypeError,
-            TypingError,
-            KeyError,
-        ):  # if can't vectorize, estimate time to pandas apply
+        except ERRORS_TO_HANDLE:  # if can't vectorize, estimate time to pandas apply
             wrapped = self._wrapped_apply(func, convert_dtype=convert_dtype, args=args, **kwds)
             timed = timeit.timeit(wrapped, number=N_REPEATS)
             sample_proc_est = timed / N_REPEATS
@@ -309,7 +310,7 @@ class DataFrameAccessor(_SwifterObject):
                     .apply(func, *args, axis=axis, raw=raw, result_type=result_type, meta=meta, **kwds)
                     .compute(scheduler=self._scheduler)
                 )
-        except (AttributeError, ValueError, TypeError, KeyError):
+        except ERRORS_TO_HANDLE:
             # if dask apply doesn't match pandas apply, fallback to pandas
             if self._progress_bar:
                 tqdm.pandas(desc=self._progress_bar_desc or "Pandas Apply")
@@ -341,13 +342,7 @@ class DataFrameAccessor(_SwifterObject):
                     error_message="Vectorized function sample does not match pandas apply sample.",
                 )
             return func(self._obj, *args, **kwds)
-        except (
-            AttributeError,
-            ValueError,
-            TypeError,
-            TypingError,
-            KeyError,
-        ):  # if can't vectorize, estimate time to pandas apply
+        except ERRORS_TO_HANDLE:  # if can't vectorize, estimate time to pandas apply
             wrapped = self._wrapped_apply(func, axis=axis, raw=raw, result_type=result_type, args=args, **kwds)
             timed = timeit.timeit(wrapped, number=N_REPEATS)
             sample_proc_est = timed / N_REPEATS
@@ -406,7 +401,7 @@ class DataFrameAccessor(_SwifterObject):
                     .applymap(func, meta=meta)
                     .compute(scheduler=self._scheduler)
                 )
-        except (AttributeError, ValueError, TypeError, KeyError):
+        except ERRORS_TO_HANDLE:
             # if dask apply doesn't match pandas apply, fallback to pandas
             if self._progress_bar:
                 tqdm.pandas(desc=self._progress_bar_desc or "Pandas Apply")
@@ -438,13 +433,7 @@ class DataFrameAccessor(_SwifterObject):
                     error_message="Vectorized function sample does not match pandas apply sample.",
                 )
             return func(self._obj)
-        except (
-            AttributeError,
-            ValueError,
-            TypeError,
-            TypingError,
-            KeyError,
-        ):  # if can't vectorize, estimate time to pandas apply
+        except ERRORS_TO_HANDLE:  # if can't vectorize, estimate time to pandas apply
             wrapped = self._wrapped_applymap(func)
             timed = timeit.timeit(wrapped, number=N_REPEATS)
             sample_proc_est = timed / N_REPEATS
@@ -489,12 +478,9 @@ class Transformation(_SwifterObject):
 
         return wrapped
 
+    @abstractmethod
     def _dask_apply(self, func, *args, **kwds):
-        if self._progress_bar:
-            with TQDMDaskProgressBar(desc=self._progress_bar_desc or "Dask Apply"):
-                return self._obj_dd.apply(func, *args, **kwds).compute(scheduler=self._scheduler)
-        else:
-            return self._obj_dd.apply(func, *args, **kwds).compute(scheduler=self._scheduler)
+        raise NotImplementedError("Transformation class does not implement _dask_apply")
 
     def apply(self, func, *args, **kwds):
         """
@@ -514,7 +500,7 @@ class Transformation(_SwifterObject):
         if est_apply_duration > self._dask_threshold:
             return self._dask_apply(func, *args, **kwds)
         else:  # use pandas
-            if self._progress_bar:
+            if self._progress_bar and hasattr(self._obj_pd, "progress_apply"):
                 tqdm.pandas(desc=self._progress_bar_desc or "Pandas Apply")
                 return self._obj_pd.progress_apply(func, *args, **kwds)
             else:
@@ -537,7 +523,7 @@ class Rolling(Transformation):
             pandas_obj, npartitions, dask_threshold, scheduler, progress_bar, progress_bar_desc, allow_dask_on_strings
         )
         self._rolling_kwds = kwds.copy()
-        self._sample_original = self._sample_pd.copy()
+        self._comparison_pd = self._obj_pd.iloc[: self._npartitions * 2]
         self._sample_pd = self._sample_pd.rolling(**kwds)
         self._obj_pd = self._obj_pd.rolling(**kwds)
         self._obj_dd = self._obj_dd.rolling(**{k: v for k, v in kwds.items() if k not in ["on", "closed"]})
@@ -547,21 +533,21 @@ class Rolling(Transformation):
             # check that the dask rolling apply matches the pandas apply
             with suppress_stdout_stderr():
                 tmp_df = (
-                    dd.from_pandas(self._sample_original, npartitions=self._npartitions)
+                    dd.from_pandas(self._comparison_pd, npartitions=self._npartitions)
                     .rolling(**{k: v for k, v in self._rolling_kwds.items() if k not in ["on", "closed"]})
                     .apply(func, *args, **kwds)
                     .compute(scheduler=self._scheduler)
                 )
-            self._validate_apply(
-                tmp_df.equals(self._sample_pd.apply(func, *args, **kwds)),
-                error_message="Dask rolling apply sample does not match pandas rolling apply sample.",
-            )
+                self._validate_apply(
+                    tmp_df.equals(self._comparison_pd.rolling(**self._rolling_kwds).apply(func, *args, **kwds)),
+                    error_message="Dask rolling apply sample does not match pandas rolling apply sample.",
+                )
             if self._progress_bar:
                 with TQDMDaskProgressBar(desc=self._progress_bar_desc or "Dask Apply"):
                     return self._obj_dd.apply(func, *args, **kwds).compute(scheduler=self._scheduler)
             else:
                 return self._obj_dd.apply(func, *args, **kwds).compute(scheduler=self._scheduler)
-        except (AttributeError, ValueError, TypeError, KeyError):
+        except ERRORS_TO_HANDLE:
             if self._progress_bar:
                 tqdm.pandas(desc=self._progress_bar_desc or "Pandas Apply")
                 return self._obj_pd.progress_apply(func, *args, **kwds)
@@ -585,7 +571,7 @@ class Resampler(Transformation):
             pandas_obj, npartitions, dask_threshold, scheduler, progress_bar, progress_bar_desc, allow_dask_on_strings
         )
         self._resampler_kwds = kwds.copy()
-        self._sample_original = self._sample_pd.copy()
+        self._comparison_pd = self._obj_pd.iloc[: self._npartitions * 2]
         self._sample_pd = self._sample_pd.resample(**kwds)
         self._obj_pd = self._obj_pd.resample(**kwds)
         # Setting dask dataframe `self._obj_dd` to None when there are 0 `self._nrows` because
@@ -601,21 +587,21 @@ class Resampler(Transformation):
             # check that the dask resampler apply matches the pandas apply
             with suppress_stdout_stderr():
                 tmp_df = (
-                    dd.from_pandas(self._sample_original, npartitions=self._npartitions)
+                    dd.from_pandas(self._comparison_pd, npartitions=self._npartitions)
                     .resample(**{k: v for k, v in self._resampler_kwds.items() if k in ["rule", "closed", "label"]})
                     .agg(func, *args, **kwds)
                     .compute(scheduler=self._scheduler)
                 )
-            self._validate_apply(
-                tmp_df.equals(self._sample_pd.apply(func, *args, **kwds)),
-                error_message="Dask resampler apply sample does not match pandas resampler apply sample.",
-            )
+                self._validate_apply(
+                    tmp_df.equals(self._comparison_pd.resample(**self._resampler_kwds).apply(func, *args, **kwds)),
+                    error_message="Dask resampler apply sample does not match pandas resampler apply sample.",
+                )
 
             if self._progress_bar:
                 with TQDMDaskProgressBar(desc=self._progress_bar_desc or "Dask Apply"):
                     return self._obj_dd.agg(func, *args, **kwds).compute(scheduler=self._scheduler)
             else:
                 return self._obj_dd.agg(func, *args, **kwds).compute(scheduler=self._scheduler)
-        except (AttributeError, ValueError, TypeError, KeyError):
+        except ERRORS_TO_HANDLE:
             # use pandas -- no progress_apply available for resampler objects
             return self._obj_pd.apply(func, *args, **kwds)